Look at the next three meals instead of just today's three meals.
If you run "menu.py -b" in the afternoon, it is likely because you are curious about what is for breakfast tomorrow, and not what was served today. I should probably make this optional so that you can just view today, or view tomorrow, or view the next three meals.